小改

+++ video_maker.py
@@ -1,4 +1,5 @@
 import os
+from PIL import Image
 from moviepy.editor import ImageSequenceClip
 from datetime import datetime
 
@@ -10,9 +11,18 @@ os.chdir(current_dir)
 folder_path = "snapshots"  # 替换为帧所在的文件夹路径
 frame_files = []
 number_of_frames = 0
+# for filename in os.listdir(folder_path):
+#     if filename.endswith(".jpg"):  # 根据文件扩展名过滤
+
+
 for filename in os.listdir(folder_path):
-    if filename.endswith(".jpg"):  # 根据文件扩展名过滤
+    if filename.endswith(".jpg"):
+        img_path = os.path.join(folder_path, filename)
+        img = Image.open(img_path)
+        img = img.resize((800, 800), Image.ANTIALIAS)
+        img.save(img_path)
         number_of_frames += 1
+
 
 frame_files = [f"{folder_path}/frame_{i}.jpg" for i in range(number_of_frames)]
 # 创建视频剪辑
@@ -24,10 +34,11 @@ clip.write_videofile(
 )
 
 # 删除原有图像
-for filename in os.listdir(folder_path):
-    try:
-        if filename.endswith(".jpg"):  # 根据文件扩展名过滤
-            file_path = os.path.join(folder_path, filename)
-            os.remove(file_path)
-    except Exception as e:
-        print("Failed to delete %s. Reason: %s" % (file_path, e))
+if input("是否删除原有图片?(Y/N)") == "Y":
+    for filename in os.listdir(folder_path):
+        try:
+            if filename.endswith(".jpg"):  # 根据文件扩展名过滤
+                file_path = os.path.join(folder_path, filename)
+                os.remove(file_path)
+        except Exception as e:
+            print("Failed to delete %s. Reason: %s" % (file_path, e))
